% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{boxplot_emergence}
\alias{boxplot_emergence}
\title{Helper function for Fig. 1-3}
\usage{
boxplot_emergence(d1, d2, ...)
}
\arguments{
\item{d1}{First list}

\item{d2}{Second list}

\item{...}{further arguments}
}
\description{
Helper function for Fig. 1-3
}
