% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{create_gamm_plot}
\alias{create_gamm_plot}
\title{Make Figures 1-3 (smoothed figures): Subfunction}
\usage{
create_gamm_plot(var, data, include_age = FALSE, modelset = list(),
  COL = c("red", "blue"), plot_sex = NULL, plot_which = NULL, ...)
}
\arguments{
\item{var}{variable to plot, e.g. time_to_sunset_min}

\item{data}{dataset}

\item{include_age}{Should age be included?}

\item{modelset}{Model for smooth line}

\item{COL}{Colors for males and females}

\item{plot_sex}{obsolete?}

\item{plot_which}{obsolete?}

\item{...}{passed on to plot.}
}
\value{
A panel of figures 1-3
}
\description{
Make Figures 1-3 (smoothed figures): Subfunction
}
