% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{gamm4.2}
\alias{gamm4.2}
\title{Wrapper function for gamm4()}
\usage{
gamm4.2(response, by_gamm, k_gamm = 40, explanatory = c(NULL), data,
  rm.intercept = FALSE, smooth_over = "rel_day")
}
\arguments{
\item{response}{Response variable}

\item{by_gamm}{"By" for the smoother}

\item{k_gamm}{"k" that si supplied to gamm4}

\item{explanatory}{explanatory variable}

\item{data}{dataset}

\item{rm.intercept}{Should the intercet be removed when calculating the model (0+ added to the model formula on the right side)}

\item{smooth_over}{Which variable to smooth over (here date or days to first egg)}
}
\value{
The output of a gamm4.
}
\description{
Wrapper function for gamm4()
}
