% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plot_descriptive_figure}
\alias{plot_descriptive_figure}
\title{Pipeline function for Figure 1-3}
\usage{
plot_descriptive_figure(modellist_yday, modellist_rel_day, d_list,
  out_list, YLAB, VAR, ylim_boxplot = c(-60, 70),
  ylim_gammplot = c(-40, 20))
}
\arguments{
\item{modellist_yday}{Modellist for "Day of year"}

\item{modellist_rel_day}{Modellist for "Days to first egg"}

\item{d_list}{Boxplot variables}

\item{out_list}{Boxplot variables}

\item{YLAB}{ylab}

\item{VAR}{which variable is plotted?}

\item{ylim_boxplot}{ylim of boxplot}

\item{ylim_gammplot}{yoim of gamm-plot}
}
\value{
Figure like Figures 1-3
}
\description{
Pipeline function for Figure 1-3
}
\details{
Uses the functions boxplot_emergence and create_gamm_plot.
}
