% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plot_env}
\alias{plot_env}
\title{Plot environmental variables (Figure S1)}
\usage{
plot_env(df, VARIABLE, ...)
}
\arguments{
\item{df}{Data}

\item{VARIABLE}{variable to plot.}

\item{...}{further parameters}
}
\value{
A figure
}
\description{
Plot environmental variables (Figure S1)
}
