% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plot_gamm_models}
\alias{plot_gamm_models}
\title{Helper function to check the fit of the model assumptions}
\usage{
plot_gamm_models(mod)
}
\arguments{
\item{mod}{gamm model to plot}
}
\value{
A figure
}
\description{
Helper function to check the fit of the model assumptions
}
