% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plot_individuals}
\alias{plot_individuals}
\title{Plot individuals}
\usage{
plot_individuals(data, i = 1, year_ID, var = "time_to_sunrise_min",
  MALE_COLOUR = "#5000FFFF", FEMALE_COLOUR = "#FF758AFF",
  plot.lines = FALSE, ...)
}
\arguments{
\item{data}{Dataset}

\item{i}{Number of the individual within data that should be plotted}

\item{year_ID}{year_ and ID pasted together}

\item{var}{variable to plot (e.g. time_to_sunset_min)}

\item{MALE_COLOUR}{colour of males}

\item{FEMALE_COLOUR}{colour of females}

\item{plot.lines}{logical. Should points within a week be connected by lines?}

\item{...}{Further arguments passed on to plot}
}
\value{
A figure such as S3-S4
}
\description{
Plot individuals
}
