% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{plot_seasonal_progress}
\alias{plot_seasonal_progress}
\title{Plots how the effect sizes vary with season (Figs. S7-10)}
\usage{
plot_seasonal_progress(tmp1, tmp2, YLIM = c(-10, 5),
  XLAB = "Onset of activity", YLAB = NULL)
}
\arguments{
\item{tmp1}{The data for day of year}

\item{tmp2}{The data for relative day}

\item{YLIM}{ylim}

\item{XLAB}{xlab}

\item{YLAB}{ylab}
}
\value{
A figure such as Figs. S7-10
}
\description{
Plots how the effect sizes vary with season (Figs. S7-10)
}
